refactor(template): introduce GenerateTabTemplate and reduce wrappers ~95%
The deep-DRY audit's biggest finding: three near-identical ~850-line
files (formatter-tabs/{json,xml,yaml}/generate-tab.tsx) sharing
~95% of their content. The 75 useState declarations for per-language
defaults, the `buildLanguageOptions` switch, the `generateCode`
dispatch switch, the stats `useEffect`, and the entire OptionsPanel
+ InputOutputSplit JSX are byte-identical across all three.

Add `GenerateTabTemplate` in `src/lib/components/template/`. The
template owns every chunk that was duplicated: ten per-language
option-state ladders (TypeScript, JavaScript, Go, Python, Rust,
Java, C#, Kotlin, Swift, PHP), the `buildLanguageOptions` switch
that satisfies the `generateCode` overload set, the language
picker, the per-language `<FormSection>` blocks in the rail, and
the final `<InputOutputSplit>` render with the dynamic
"Generated Code (Java)" header. Per-format variation is funneled
through eight props: `parseInput`, `validateInput`,
`inputEditorMode`, `inputTitle`, `inputPlaceholder`, `emptyTitle`,
`emptyClearValue`, and `extraOptions` (the JSON wrapper's
`<JsonFormatSection>` injection).

Replace the three wrappers with thin ~30-40-line files that hand in
the per-format pieces:

* `formatter-tabs/json/generate-tab.tsx` — threads the project-wide
  `useJsonFormatterOptions().inputFormat` through `parseJson` /
  `validateJson`, injects `<JsonFormatSection>`, opts into
  `emptyClearValue='{}'`.
* `formatter-tabs/xml/generate-tab.tsx` — routes through
  `xmlToJsonObject` and validates via the `DOMParser` `parsererror`
  check.
* `formatter-tabs/yaml/generate-tab.tsx` — routes through
  `yaml.parse` for both parse and validate (the library throws on
  invalid input, which the template catches).

Net change: 3 files lose 2,501 lines, 1 new template file gains
~870 lines, the template barrel gains 3 lines. Real net reduction
is about **−1,560 lines** — the single largest win in the polish
series and removes the maintenance hazard where a fix in one
wrapper had to be ported to the other two.

All 141 existing tests pass.
