+++ src/lib/components/template/index.ts
@@ -4,5 +4,8 @@ export type { CompareTabProps } from './compare-tab';
 export { ConvertTab } from './convert-tab';
 export type { ConvertTabProps } from './convert-tab';
 
+export { GenerateTabTemplate } from './generate-tab';
+export type { GenerateTabTemplateProps, TabStats as GenerateTabStats } from './generate-tab';
+
 export { TabbedFormatterPage } from './tabbed-formatter-page';
 export type { TabbedFormatterPageProps, TabContentProps } from './tabbed-formatter-page';
